Refactor WebSocket handling: simplify WebSocketActor by removing direct service dependencies and move word count logic to WebSocketService for better separation of concerns
